Added tests for new burn mechanism logic and separated test files
